llm_mentor/requirements.txt
openai>=1.30.0,<2.0.0

langchain==0.2.17
langchain-community==0.2.17
langchain-openai==0.1.7

chromadb==0.5.5

tiktoken==0.7.0
pandas==2.2.2

beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.32.3

gradio==4.44.0
python-dotenv==1.0.1
